+++ R/assert_functions.R
@@ -25,6 +25,33 @@ function_expects_n_arguments_advanced <- function(x, n, dots = c("throw_error","
 }
 
 
+# Expect function has arguments
+# @param
+# @param required the names of parameters the function must support (character)
+#
+# @returns TRUE if the function contains the expected parameters, otherwise returns a string (the error message)
+#
+function_expects_advanced <- function(x, required){
+  if(!is.function(x)) {
+    value_class <- toString(class(x))
+    return(paste0("{.strong '{arg_name}'} must be a function, not a ", value_class))
+  }
+
+  declared_args <- setdiff(func_arg_names(x), "...")
+  if(!is_subset(required, declared_args)){
+    missing_args <- setopts_exlusive_to_first(required, declared_args)
+    missing_count <- length(missing_args)
+    missing_args <- paste0("`", paste(missing_args, collapse = "`, `"), "`")
+    return(paste0("Function '{arg_name}' is missing the following parameter",
+                  if(missing_count == 1) "" else "s",
+                  " in its signature: {.strong ", missing_args, "}"
+    ))
+  }
+
+  return(TRUE)
+}
+
+
 # Assertions --------------------------------------------------------------
 
 #' Assert function expects n arguments
@@ -42,3 +69,28 @@ function_expects_n_arguments_advanced <- function(x, n, dots = c("throw_error","
 #'
 #' @export
 assert_function_expects_n_arguments <- assert_create(func = function_expects_n_arguments_advanced)
+
+#' Assert function expects specific parameter names
+#'
+#' Assert that a function signature includes required set of parameter names in its
+#' formal argument list, regardless of whether those parameters have default
+#' values. The `...` argument is ignored.
+#'
+#' @param x a function to check for required parameter names
+#' @param required a character vector of parameter names that must appear in
+#'   the function signature (order does not matter)
+#' @inheritParams common_roxygen_params
+#'
+#' @return invisible(TRUE) if function `x` declares all required parameters,
+#'   otherwise aborts with the error message specified by `msg`
+#'
+#' @examples
+#' my_fun <- function(x, y = 1, ...) x + y
+#' assert_function_expects(my_fun, c("x", "y"))
+#'
+#' try({
+#'   assert_function_expects(my_fun, c("x", "z"))
+#' })
+#'
+#' @export
+assert_function_expects <- assert_create(func = function_expects_advanced)